Clean up testing a bit, add COC and CONTRIBUTING
