+++ .gitignore
@@ -71,4 +71,4 @@ lib/extensions
 # Autogenerated content
 lib/grammars
 lib/themes
-test/integration/report.html
+test/integration/viewer/report.html
