+++ test/integration/baselines/highlight-directive-comments.html
