+++ test/integration/baselines/npm-extension.html
