+++ test/integration/baselines/vsix.html
