--- test/integration/cases/github-zip.js
@@ -1,6 +1,0 @@
-const path = require('path');
-
-module.exports = {
-  theme: 'Oceanic Plus',
-  extensions: [path.resolve(__dirname, '../../fixtures/oceanic-plus-master.zip')]
-};
