+++ test/integration/cases/github-zip/options.js
@@ -1,0 +1,6 @@
+const path = require('path');
+
+module.exports = {
+  theme: 'Oceanic Plus',
+  extensions: [path.resolve(__dirname, '../../../fixtures/oceanic-plus-master.zip')]
+};
