+++ test/integration/cases/npm-extension/options.js
