+++ test/integration/cases/npm-extension/test.md
