+++ test/integration/cases/parent-selector/options.js
