--- test/integration/cases/vsix.js
@@ -1,6 +1,0 @@
-const path = require('path');
-
-module.exports = {
-  theme: 'One Dark Pro',
-  extensions: [path.resolve(__dirname, '../../fixtures/one-dark-pro.vsix')]
-};
