+++ test/integration/cases/vsix/options.js
@@ -1,0 +1,6 @@
+const path = require('path');
+
+module.exports = {
+  theme: 'One Dark Pro',
+  extensions: [path.resolve(__dirname, '../../../fixtures/one-dark-pro.vsix')]
+};
