--- test/integration/options.js
@@ -1,5 +1,0 @@
-const path = require('path');
-
-module.exports = {
-  extensionDataDirectory: path.resolve(__dirname, 'extensions'),
-};
