+++ test/integration/viewer/render.js
