+++ test/setup.js
@@ -1,0 +1,4 @@
+module.exports = async (config) => {
+  // console.log(config);
+  process.env.JEST_CONFIG = JSON.stringify(config);
+};
